Learn variables, expressions and statements
In this chapter of the course Python for everybody, I learned about variables, expressions and statements.
I wrote 3 small codes using *input*, print()fonction.
A focus on types(integer, string, float, ), values, operators, operands, rules of precedence (PEMDAS), and #comments.
So............
What...is the airspeed velocity of an unladen swallow?