Feat : files in the secondarary memory
In this chapter, we start to work with Secondary Memory (or files)
File handle : open, close, read, write
Reading files, Searching through a file, Letting the user choose the file name
Using try, except, and open.
We could close the files which we open for read as well, but we can be a little
sloppy if we are only opening a few files since Python makes sure that all open
files are closed when the program ends. When we are writing files, we want to
explicitly close the files so as to leave nothing to chance.
Glossary
catch
To prevent an exception from terminating a program using the try and except statements.
newline
A special character used in files and strings to indicate the end of a line.
Pythonic
A technique that works elegantly in Python. “Using try and except is the Pythonic way to recover from missing files”.
Quality Assurance
A person or team focused on insuring the overall quality of a software product. QA is often involved in testing a product and identifying problems before the product is released.
text file
A sequence of characters stored in permanent storage like a hard drive.
Quality Assurance team whose very job it is to do the craziest things
possible in an attempt to break the software that the programmer has created.