Feat: changed files name and completed ex_07_02
A very rewarding exercise to complete. Python started to pay off.
Exercise 7.2 Write a program that prompts for a file name, then opens that file and
reads through the file, looking for lines of the form: X-DSPAM-Confidence 0.8475
Count these lines and extract the floating point values from each of the lines
and compute the average of those values and produce an output as shown
below.